Add a GHA workflow for automating agent version updates.
This adds a manually triggered workflow for generating PRs to update the
version of the Netdata Agent provided by the Helm chart.

Short-term, this will be manually triggered during the release process,
and will simply serve to speed things up a bit.

Longer-term, this will be automatically triggered by the release
automation code in the agent repo once the Docker images have been
published, reducing the amount of manual involvement required from the
release manager.