Delete uncommitted release tag
Tag release step happens before committing the new version of the helmchart in Chart.yaml. This can lead to inconsistencies and cause following release actions to fail since "next" release tag has been pushed to remote but current_version on Chart.yaml is still the previous, so it always try to push an existing version.

This commit adds a check if commit step failed and deletes the previously pushed release tag